Handled exceptions for both cinema and movie ratings Added The Avg for cinema rating

+++ src/main/java/com/roland/movietheater_jdbc/repository/RatingMovieRepository/RatingMovieRepository.java
@@ -54,7 +54,7 @@ public class RatingMovieRepository implements IRatingMovieRepository {
     @Override
     public double getAverageRatingForMovieForUser(int movieId) throws FailedToRateMovie {
         try {
-            return  jdbcTemplate.queryForObject(SQL_STATEMENT_TO_FIND_THE_AVG_RATING_FOR_MOVIE,Integer.class,movieId);
+            return  jdbcTemplate.queryForObject(SQL_STATEMENT_TO_FIND_THE_AVG_RATING_FOR_MOVIE,Double.class,movieId);
         } catch (DataAccessException e) {
             throw new FailedToRateMovie("Movie Not Found !");
         }
